requirements.txt
feedparser==5.2.1
python-telegram-bot==3.2.0
psycopg2==2.6.1
